# feat: add 'pretty' parameter to optimize JSON serialization performance